Supply the match object used to the plugin object
If the user language does not exactly match the language associated with
the regex used to select a plugin action, the code responsible for
recreating the match would fail to find the regex. To avoid this instead
of redoing the match inside the plugin, the match object is passed in
by PluginManager.